requirements.txt
# Core Dependencies
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20

# AI/ML
ultralytics==8.3.120
torch==2.7.0
torchvision==0.22.0

# Image Processing
opencv-python==4.11.0.86
pillow==11.2.1
numpy==2.2.5

# Utilities
pydantic==2.11.3
Jinja2==3.1.6
requests==2.32.3









# annotated-types==0.7.0
# anyio==4.9.0
# certifi==2025.4.26
# charset-normalizer==3.4.1
# click==8.1.8
# contourpy==1.3.2
# cycler==0.12.1
# fastapi==0.115.12
# filelock==3.18.0
# fonttools==4.57.0
# fsspec==2025.3.2
# gitdb==4.0.12
# GitPython==3.1.44
# h11==0.16.0
# idna==3.10
# Jinja2==3.1.6
# kiwisolver==1.4.8
# MarkupSafe==3.0.2
# matplotlib==3.10.1
# mpmath==1.3.0
# networkx==3.4.2
# numpy==2.2.5
# nvidia-cublas-cu12==12.6.4.1
# nvidia-cuda-cupti-cu12==12.6.80
# nvidia-cuda-nvrtc-cu12==12.6.77
# nvidia-cuda-runtime-cu12==12.6.77
# nvidia-cudnn-cu12==9.5.1.17
# nvidia-cufft-cu12==11.3.0.4
# # nvidia-cufile-cu12==1.11.1.6    for linux
# nvidia-curand-cu12==10.3.7.77
# nvidia-cusolver-cu12==11.7.1.2
# nvidia-cusparse-cu12==12.5.4.2
# nvidia-cusparselt-cu12==0.6.3
# # nvidia-nccl-cu12==2.26.2       for linux
# nvidia-nvjitlink-cu12==12.6.85
# nvidia-nvtx-cu12==12.6.77
# opencv-python==4.11.0.86
# packaging==25.0
# pandas==2.2.3
# pillow==11.2.1
# psutil==7.0.0
# py-cpuinfo==9.0.0
# pydantic==2.11.3
# pydantic_core==2.33.1
# pyparsing==3.2.3
# python-dateutil==2.9.0.post0
# python-multipart==0.0.20
# pytz==2025.2
# PyYAML==6.0.2
# requests==2.32.3
# scipy==1.15.2
# seaborn==0.13.2
# setuptools==80.0.0
# six==1.17.0
# smmap==5.0.2
# sniffio==1.3.1
# starlette==0.46.2
# sympy==1.14.0
# torch==2.7.0
# torchaudio==2.7.0
# torchvision==0.22.0
# tqdm==4.67.1
# triton==3.3.0
# typing-inspection==0.4.0
# typing_extensions==4.13.2
# tzdata==2025.2
# ultralytics==8.3.120
# ultralytics-thop==2.0.14
# urllib3==2.4.0
# uvicorn==0.34.2
